utils: drop tests_import
Remove dead code in utils.tests_import which was used with the
previous API.  The new API for imporing test results uses
utils.kci_test instead.  Move code from tests_import to kci_test and
drop the utils.tests_import module.
